spi-flasher: Change start value of QSPI1_PADCFG register
CLE bit is a test bit and should not be set to 1, as it may lead to a
voltage drop. Initial setting of the QSPI1 pads should not use 1.8V,
as this may damage the pads if 3.3V is actually used.

At start the value of QSPI1_PADCFG register has been changed from
0x7 to 0x1, CLE bit and v18 bit is set to 0.

Change-Id: Ic162f01ddedcf6955f0bcdc6e2dc3b1bbadf4941